+++ tsdb/tsm1/file_store.go
@@ -1210,7 +1210,7 @@ func (f *FileStore) locations(key []byte, t int64, ascending bool) []*location {
 
 // CreateSnapshot creates hardlinks for all tsm and tombstone files
 // in the path provided.
-func (f *FileStore) CreateSnapshot(ctx context.Context) (string, error) {
+func (f *FileStore) CreateSnapshot(ctx context.Context) (backupID int, backupDirFullPath string, err error) {
 	span, _ := tracing.StartSpanFromContext(ctx)
 	defer span.Finish()
 
@@ -1230,30 +1230,35 @@ func (f *FileStore) CreateSnapshot(ctx context.Context) (string, error) {
 	// increment and keep track of the current temp dir for when we drop the lock.
 	// this ensures we are the only writer to the directory.
 	f.currentTempDirID += 1
-	tmpPath := fmt.Sprintf("%d.%s", f.currentTempDirID, TmpTSMFileExtension)
-	tmpPath = filepath.Join(f.dir, tmpPath)
+	backupID = f.currentTempDirID
 	f.mu.Unlock()
+
+	backupDirFullPath = f.InternalBackupPath(backupID)
 
 	// create the tmp directory and add the hard links. there is no longer any shared
 	// mutable state.
-	err := os.Mkdir(tmpPath, 0777)
+	err = os.Mkdir(backupDirFullPath, 0777)
 	if err != nil {
-		return "", err
+		return 0, "", err
 	}
 	for _, tsmf := range files {
-		newpath := filepath.Join(tmpPath, filepath.Base(tsmf.Path()))
+		newpath := filepath.Join(backupDirFullPath, filepath.Base(tsmf.Path()))
 		if err := os.Link(tsmf.Path(), newpath); err != nil {
-			return "", fmt.Errorf("error creating tsm hard link: %q", err)
+			return 0, "", fmt.Errorf("error creating tsm hard link: %q", err)
 		}
 		for _, tf := range tsmf.TombstoneFiles() {
-			newpath := filepath.Join(tmpPath, filepath.Base(tf.Path))
+			newpath := filepath.Join(backupDirFullPath, filepath.Base(tf.Path))
 			if err := os.Link(tf.Path, newpath); err != nil {
-				return "", fmt.Errorf("error creating tombstone hard link: %q", err)
+				return 0, "", fmt.Errorf("error creating tombstone hard link: %q", err)
 			}
 		}
 	}
 
-	return tmpPath, nil
+	return backupID, backupDirFullPath, nil
+}
+
+func (f *FileStore) InternalBackupPath(backupID int) string {
+	return filepath.Join(f.dir, fmt.Sprintf("%d.%s", backupID, TmpTSMFileExtension))
 }
 
 // MeasurementStats returns the sum of all measurement stats within the store.